+++ test/pathological/file_stores_test.clj
@@ -1,0 +1,3 @@
+(ns pathological.file-stores-test
+  (:require
+    [clojure.test :refer :all]))